# docs/contracts/BASIC_CURRENT_DEFINITION_OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DCP-BASIC-NO-AUDIT-VERSION-001 bounded API supplement",
    "version": "2026-10-09",
    "description": "Approved supplement to FINAL BASELINE v1.0.23. Only listed basic controls/process bindings supersede prior contracts. Historical evidence and regulated controls remain. Fully qualified schema IDs prevent Create/Update collisions without changing HTTP payload fields. DCP-EBR-TEMPLATE-NAME-001 adds required create templateName and controlled draft name editing; historical canonical content remains unchanged."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/ebr/templates/{id}/versions": {
      "post": {
        "tags": [
          "ebr-controller"
        ],
        "operationId": "version",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.hospital.mes.ebr.domain.EbrCommands.VersionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "com.hospital.mes.ebr.domain.EbrCommands.VersionCreate": {
        "type": "object",
        "properties": {
          "versionNo": {
            "type": "integer",
            "format": "int64"
          },
          "reason": {
            "type": "string"
          },
          "copyDefinition": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}